projects/claude-pdf-summarizer/requirements.txt
fastapi
uvicorn[standard]
anthropic
python-dotenv
PyPDF2
python-multipart
reportlab
requests
